Update @github/copilot to 1.0.51-2 (#1342)
* Update @github/copilot to 1.0.51-2

- Updated nodejs and test harness dependencies
- Re-ran code generators
- Formatted generated code

* Fix Go E2E test: use %f for float64 TotalNanoAiu fields

The TotalNanoAiu field was changed from int to *float64 in the generated
types, but the test format strings still used %d. This caused go vet to
fail on all three Go CI legs.


---------

+++ go/rpc/zsession_events.go
@@ -1050,9 +1050,9 @@ type SessionShutdownData struct {
 	// Cumulative time spent in API calls during the session, in milliseconds
 	TotalAPIDurationMs int64 `json:"totalApiDurationMs"`
 	// Session-wide accumulated nano-AI units cost
-	TotalNanoAiu *int64 `json:"totalNanoAiu,omitempty"`
+	TotalNanoAiu *float64 `json:"totalNanoAiu,omitempty"`
 	// Total number of premium API requests used during the session
-	TotalPremiumRequests int64 `json:"totalPremiumRequests"`
+	TotalPremiumRequests float64 `json:"totalPremiumRequests"`
 }
 
 func (*SessionShutdownData) sessionEventData()      {}
@@ -1465,7 +1465,7 @@ type AssistantUsageCopilotUsage struct {
 	// Itemized token usage breakdown
 	TokenDetails []AssistantUsageCopilotUsageTokenDetail `json:"tokenDetails"`
 	// Total cost in nano-AI units for this request
-	TotalNanoAiu int64 `json:"totalNanoAiu"`
+	TotalNanoAiu float64 `json:"totalNanoAiu"`
 }
 
 // Token usage detail for a single billing category
@@ -1537,7 +1537,7 @@ type CompactionCompleteCompactionTokensUsedCopilotUsage struct {
 	// Itemized token usage breakdown
 	TokenDetails []CompactionCompleteCompactionTokensUsedCopilotUsageTokenDetail `json:"tokenDetails"`
 	// Total cost in nano-AI units for this request
-	TotalNanoAiu int64 `json:"totalNanoAiu"`
+	TotalNanoAiu float64 `json:"totalNanoAiu"`
 }
 
 // Token usage detail for a single billing category
@@ -2225,7 +2225,7 @@ type ShutdownModelMetric struct {
 	// Token count details per type
 	TokenDetails map[string]ShutdownModelMetricTokenDetail `json:"tokenDetails,omitempty"`
 	// Accumulated nano-AI units cost for this model
-	TotalNanoAiu *int64 `json:"totalNanoAiu,omitempty"`
+	TotalNanoAiu *float64 `json:"totalNanoAiu,omitempty"`
 	// Token usage breakdown
 	Usage ShutdownModelMetricUsage `json:"usage"`
 }
